Correct return type

+++ src/core/functions/scalar/weakly_connected_component.cpp
@@ -177,7 +177,7 @@ void CoreScalarFunctions::RegisterWeaklyConnectedComponentScalarFunction(
       ScalarFunction(
           "weakly_connected_component",
           {LogicalType::INTEGER, LogicalType::BIGINT},
-          LogicalType::FLOAT, WeaklyConnectedComponentFunction,
+          LogicalType::BIGINT, WeaklyConnectedComponentFunction,
           WeaklyConnectedComponentFunctionData::WeaklyConnectedComponentBind));
 }
